+++ test/Builder/ShlinkClientBuilderTest.php
@@ -5,8 +5,6 @@ declare(strict_types=1);
 namespace ShlinkioTest\Shlink\SDK\Builder;
 
 use PHPUnit\Framework\TestCase;
-use Prophecy\PhpUnit\ProphecyTrait;
-use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestFactoryInterface;
 use Psr\Http\Message\StreamFactoryInterface;
@@ -17,24 +15,16 @@ use Shlinkio\Shlink\SDK\Config\ShlinkConfigInterface;
 class ShlinkClientBuilderTest extends TestCase
 {
     use ClientBuilderMethodsProviderTrait;
-    use ProphecyTrait;
 
     private ShlinkClientBuilder $builder;
-    private ObjectProphecy $client;
-    private ObjectProphecy $requestFactory;
-    private ObjectProphecy $streamFactory;
     private ShlinkConfigInterface $config;
 
     public function setUp(): void
     {
-        $this->client = $this->prophesize(ClientInterface::class);
-        $this->requestFactory = $this->prophesize(RequestFactoryInterface::class);
-        $this->streamFactory = $this->prophesize(StreamFactoryInterface::class);
-
         $this->builder = new ShlinkClientBuilder(
-            $this->client->reveal(),
-            $this->requestFactory->reveal(),
-            $this->streamFactory->reveal(),
+            $this->createMock(ClientInterface::class),
+            $this->createMock(RequestFactoryInterface::class),
+            $this->createMock(StreamFactoryInterface::class),
         );
         $this->config = ShlinkConfig::fromBaseUrlAndApiKey('foo', 'bar');
     }