+++ test/Tags/TagsClientTest.php
@@ -5,10 +5,8 @@ declare(strict_types=1);
 namespace ShlinkioTest\Shlink\SDK\Tags;
 
 use DMS\PHPUnitExtensions\ArraySubset\ArraySubsetAsserts;
+use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
-use Prophecy\Argument;
-use Prophecy\PhpUnit\ProphecyTrait;
-use Prophecy\Prophecy\ObjectProphecy;
 use Shlinkio\Shlink\SDK\Exception\InvalidDataException;
 use Shlinkio\Shlink\SDK\Http\Exception\HttpException;
 use Shlinkio\Shlink\SDK\Http\HttpClientInterface;
@@ -19,19 +17,19 @@ use Shlinkio\Shlink\SDK\Tags\Model\TagRenaming;
 use Shlinkio\Shlink\SDK\Tags\Model\TagsFilter;
 use Shlinkio\Shlink\SDK\Tags\Model\TagsListOrderField;
 use Shlinkio\Shlink\SDK\Tags\TagsClient;
+use Throwable;
 
 class TagsClientTest extends TestCase
 {
     use ArraySubsetAsserts;
-    use ProphecyTrait;
 
     private TagsClient $tagsClient;
-    private ObjectProphecy $httpClient;
+    private MockObject & HttpClientInterface $httpClient;
 
     public function setUp(): void
     {
-        $this->httpClient = $this->prophesize(HttpClientInterface::class);
-        $this->tagsClient = new TagsClient($this->httpClient->reveal());
+        $this->httpClient = $this->createMock(HttpClientInterface::class);
+        $this->tagsClient = new TagsClient($this->httpClient);
     }
 
     /** @test */
@@ -59,7 +57,7 @@ class TagsClientTest extends TestCase
     public function listTagsWithStatsReturnsExpectedResponse(): void
     {
         $this->assertListTags(
-            ['/tags/stats', Argument::type('array')],
+            ['/tags/stats', $this->isType('array')],
             [[], [], [], [], []],
             function (): array {
                 $iterable = $this->tagsClient->listTagsWithStats();
@@ -80,7 +78,7 @@ class TagsClientTest extends TestCase
         $filter = TagsFilter::create()->searchingBy('foo')->orderingAscBy(TagsListOrderField::TAG);
         $test = $this;
         $this->assertListTags(
-            ['/tags/stats', Argument::that(function (array $arg) use ($filter, $test) {
+            ['/tags/stats', $this->callback(function (array $arg) use ($filter, $test) {
                 $test->assertArraySubset($filter->toArray(), $arg);
                 return true;
             })],
@@ -100,7 +98,7 @@ class TagsClientTest extends TestCase
 
     private function assertListTags(array $expectedArgs, array $expectedData, callable $listTags): void
     {
-        $get = $this->httpClient->getFromShlink(...$expectedArgs)->willReturn([
+        $this->httpClient->expects($this->once())->method('getFromShlink')->with(...$expectedArgs)->willReturn([
             'tags' => [
                 'data' => $expectedData,
             ],
@@ -109,29 +107,25 @@ class TagsClientTest extends TestCase
         $result = $listTags();
 
         self::assertEquals($expectedData, $result);
-        $get->shouldHaveBeenCalledOnce();
     }
 
     /** @test */
     public function renameTagCallsApi(): void
     {
         $renaming = TagRenaming::fromOldNameAndNewName('foo', 'bar');
-        $call = $this->httpClient->callShlinkWithBody('/tags', 'PUT', $renaming);
+        $this->httpClient->expects($this->once())->method('callShlinkWithBody')->with('/tags', 'PUT', $renaming);
 
         $this->tagsClient->renameTag($renaming);
-
-        $call->shouldHaveBeenCalledOnce();
     }
 
     /**
+     * @param class-string<Throwable> $expectedException
      * @test
      * @dataProvider provideRenameExceptions
      */
     public function renameTagThrowsProperExceptionOnError(HttpException $original, string $expectedException): void
     {
-        $call = $this->httpClient->callShlinkWithBody(Argument::cetera())->willThrow($original);
-
-        $call->shouldBeCalledOnce();
+        $this->httpClient->expects($this->once())->method('callShlinkWithBody')->willThrowException($original);
         $this->expectException($expectedException);
 
         $this->tagsClient->renameTag(TagRenaming::fromOldNameAndNewName('', ''));
@@ -163,22 +157,24 @@ class TagsClientTest extends TestCase
     public function deleteTagsCallsApi(): void
     {
         $tags = ['foo', 'bar', 'baz'];
-        $call = $this->httpClient->callShlinkWithBody('/tags', 'DELETE', [], ['tags' => $tags]);
+        $this->httpClient->expects($this->once())->method('callShlinkWithBody')->with(
+            '/tags',
+            'DELETE',
+            [],
+            ['tags' => $tags],
+        );
 
         $this->tagsClient->deleteTags(...$tags);
-
-        $call->shouldHaveBeenCalledOnce();
     }
 
     /**
+     * @param class-string<Throwable> $expectedException
      * @test
      * @dataProvider provideDeleteExceptions
      */
     public function deleteTagsThrowsProperExceptionOnError(HttpException $original, string $expectedException): void
     {
-        $call = $this->httpClient->callShlinkWithBody(Argument::cetera())->willThrow($original);
-
-        $call->shouldBeCalledOnce();
+        $this->httpClient->expects($this->once())->method('callShlinkWithBody')->willThrowException($original);
         $this->expectException($expectedException);
 
         $this->tagsClient->deleteTags('foo');